requirements.txt
hyperjson
orjson
python-rapidjson
ujson
tabulate